Calculate empty cells in row (not tested)

+++ row_solver.hpp
@@ -6,4 +6,4 @@
 /// @param cells      - row, vector of cells with current state
 /// @param block_list - vector of pair (color number, block length)
 /// @return  number of cells with changed state
-std::vector<int> calculate_row(std::vector<Cell> &cells, std::vector<std::pair<int, int>> &blocks);
+std::vector<size_t> calculate_row(std::vector<Cell> &cells, std::vector<std::pair<int, int>> &blocks);
